+++ node/core/dispute-coordinator/src/initialized.rs
@@ -52,9 +52,9 @@ use polkadot_primitives::v2::{
 use crate::{
 	error::{log_error, Error, FatalError, FatalResult, JfyiError, JfyiResult, Result},
 	metrics::Metrics,
-	real::{ordering::get_finalized_block_number, DisputeCoordinatorSubsystem},
+	ordering::get_finalized_block_number,
 	status::{get_active_with_status, Clock, DisputeStatus, Timestamp},
-	LOG_TARGET,
+	DisputeCoordinatorSubsystem, LOG_TARGET,
 };
 
 use super::{
